Demand Forecasting 1/requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
xgboost
prophet
statsmodels
tensorflow
google-generativeai